// actyondatafront/src/app/shared/models/cliente-model.ts
export class clienteModel {
    cliente_id;
    CONT_ID: Number;
    NOME: String;
    CPF: Number;
    RG: Number;
    IDENTIFICADOR_ID: String;
    DATA_NASCIMENTO: Date;
    SEXO: String;
    CEP: Number;
    ENDERECO: String;
    NUMERO: Number;
    COMPLEMENTO: String;
    BAIRRO: String;
    CIDADE: String;
    UF: String;
    NOME_CONJUGUE: String;
    NOME_PAI: String;
    NOME_MAE: String;
    ESTADO_CIVIL: String;
}